requirements.txt
psutil~=5.9.8
scapy~=2.5.0
python-pytun~=2.4.1
